fixed up the slurm generation '

+++ Fastsim/Slurm/run_pat.sh
@@ -1,16 +1,22 @@
 #!/bin/bash
 
 run_num=$1
-infilename=
-outfilename=miniAOD.${run_num}.root
+infilename=AODSIM/AODSIM.${run_num}.root
+outfilename=miniAOD/miniAOD.${run_num}.root
 n_events=$2
+
+if [ ! -f $infilename ] 
+then
+    echo "$infilename not found in AODSIM/"
+    exit
+fi
 
 cp pat_default_cfg.py logfiles/patc_${run_num}_cfg.py
 cd logfiles
 
 sed -i -e s/NUMBER_EVENT/$n_events/g patc_${run_num}_cfg.py
-sed -i -e s/INFILE/${filename}/g patc_${run_num}_cfg.py
-sed -i -e s/OUTFILE/${outfilename}/g patc_${run_num}_cfg.py
+sed -i -e s@INFILE@${infilename}@g patc_${run_num}_cfg.py
+sed -i -e s@OUTFILE@${outfilename}@g patc_${run_num}_cfg.py
 
 touch patlog.${run_num}.log
 cd ..
